Move statusMessages sync-schema to separate file
